moved some files out of django project repo
