+++ review_crawler.py
@@ -1,28 +1,28 @@
-"""Test of Beautiful Soup."""
+"""Format requests from GoodReads website with Beautiful Soup."""
 
 from bs4 import BeautifulSoup
-from datetime import datetime
 import requests
-import re
 import pdb
 
 
-request = requests.get('https://www.goodreads.com/book/show/28259132-chaos-monkeys#other_reviews')
+def review_maker():
+    """Parse request from GoodReads website with BS."""
+    review_texts = []
+    star_num_list = []
 
-#request.content is where it's at
+    for i in range(1, 9):
+        request = requests.get('https://www.rottentomatoes.com/m/star_wars_the_last_jedi/reviews/?page=%d&type=user&sort=' % i)
+        if request.status_code == 200:
 
-##page = urllib2.urlopen(request)
-soup = BeautifulSoup(request.content, 'html.parser')
+            soup = BeautifulSoup(request.content, 'html.parser')
 
-review_texts = []
-stars_text = []
+            all_reviews = soup.find_all(class_="user_review")
 
-all_reviews = soup.find_all(class_="friendReviews")
+            for text in all_reviews:
+                review = text.get_text()
+                review_texts.append(review)
 
-for text in all_reviews:
-    review = text.find_all('span', id=lambda val: val and val.startswith('freeTextContainer'))[0].string
-    review_texts.append(review)
-
-for star in all_reviews:
-    number_stars = len(star.find(class_='staticStars').find_all(class_='p10'))
-    stars_text.append(star)
+            for star in all_reviews:
+                num_stars = star.find(class_='scoreWrapper').span['class'][0]
+                star_num_list.append(num_stars)
+    pdb.set_trace()
