+++ reviews_to_improve/reviews_to_improve/settings.py
@@ -24,7 +24,7 @@ SECRET_KEY = os.environ.get('SECRET_KEY', '')
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = os.environ.get('DEBUG', '')
 
-ALLOWED_HOSTS = ['ec2-54-146-255-108.compute-1.amazonaws.com', '127.0.0.1', 'localhost']
+# ALLOWED_HOSTS = ['ec2-54-146-255-108.compute-1.amazonaws.com', '127.0.0.1', 'localhost']
 
 
 # Application definition
